# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.ErrorResponse:
    description: Error response returned when a request fails
    properties:
      error:
        example: additional error details
        type: string
      message:
        example: Invalid credentials
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@actalog.com
    name: ActaLog Support
    url: https://github.com/johnzastrow/actalog/issues
  description: ActaLog is a mobile-first CrossFit workout tracker API. It provides
    endpoints for user authentication, workout logging, movement tracking, personal
    records, and administrative functions.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://actalog.com/terms/
  title: ActaLog API
  version: 0.17.0
paths:
...
  /export/movements:
    get:
      consumes:
      - application/json
      description: Export movements to CSV or JSON format with optional filters
      parameters:
      - description: Include standard movements (default true)
        in: query
        name: include_standard
        type: boolean
      - description: Include custom movements (default true)
        in: query
        name: include_custom
        type: boolean
      - description: 'Export format: csv or json (default csv)'
        in: query
        name: format
        type: string
      produces:
      - text/csv
      - application/json
      responses:
        "200":
          description: Exported file (CSV or JSON)
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export movements
      tags:
      - import-export
  /export/user-workouts:
    get:
      consumes:
      - application/json
      description: Export user's logged workouts to CSV or JSON format with optional
        date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: 'Export format: csv or json (default json)'
        in: query
        name: format
        type: string
      produces:
      - text/csv
      - application/json
      responses:
        "200":
          description: Exported file (CSV or JSON)
          schema:
            type: file
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export user workouts
      tags:
      - import-export
  /export/wods:
    get:
      consumes:
      - application/json
      description: Export WODs to CSV or JSON format with optional filters
      parameters:
      - description: Include standard WODs (default true)
        in: query
        name: include_standard
        type: boolean
      - description: Include custom WODs (default true)
        in: query
        name: include_custom
        type: boolean
      - description: 'Export format: csv or json (default csv)'
        in: query
        name: format
        type: string
      produces:
      - text/csv
      - application/json
      responses:
        "200":
          description: Exported file (CSV or JSON)
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export WODs
      tags:
      - import-export
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"